Use erb template for rizzo-next header
When user registers we want to set proper source in Sailthru. If he
comes from Luna it should be `luna`, from Thorn Tree it should be
`community`, etc.

This can't be achieved with static generated HTML, not in the current
setup when file is not aware of the server role.

This will require additional manual step to update the template.
An `npm postinstall` instructions how to update the nav bar have been
added:

```
01:28:56 {ru-sailthru-source} ~/projects/rizzo$ npm install

> rizzo@0.1.2 postinstall /home/rob/projects/rizzo
> node rizzo-next-post-install.js

********************************************************
* To update global-header adjust the file:             *
*                                                      *
* app/views/layouts/_rizzo_next_global_header.erb.html *
*                                                      *
* to match the file:                                   *
*                                                      *
* node_modules/rizzo-next/dist/header.html             *
********************************************************
```

This is not ideal solution and works only in legazy rizzo applications,
but neither is the connection between rizzo-next and rizzo. Idealy we
would like to decouple both applications, make jenkins aware of app
identity and configure webpack to generate the file properly. However
right now there is no way to do this, and it's quite far road ahead of
us, so suboptimal solution has to do.